Some things shouldnd't happen on built-in

+++ pipelines/libs/global/vars/updateCiNode.groovy
@@ -7,19 +7,27 @@ def call(String agentName, Map info)
     println("Running updateCiNode on ${agentName}")
 
     node("${agentName}") {
-	// Update Jenkins node labels from getNodes()
-	// - this preserves 'down'
-	updateLabels(info)
 
-	// Clear it out
+	// Some things don't need to run on the 'built-in' Jenkins node
+	if (agentName != 'built-in') {
+	    // Update Jenkins node labels from getNodes()
+	    // - this preserves 'down'
+	    updateLabels(info)
+	    if (params.reinstall == '1') {
+		sh '''
+                     rm -f /bin/citbash
+                   '''
+	    }
+	}
+
+	// Clear out ci-tools
 	if (params.reinstall == '1') {
 	    sh '''
-                 rm -f /bin/citbash
                  rm -rf $HOME/ci-tools
                '''
 	}
-
-	// Update it
+	
+	// Update them
 	sh '''
              if [ -d $HOME/ci-tools ]; then
                cd $HOME/ci-tools
@@ -28,10 +36,15 @@ def call(String agentName, Map info)
                 cd $HOME
                 git clone https://github.com/kronosnet/ci-tools.git
              fi
-
-             if [ ! -f /bin/citbash ]; then
-               ln -sf `which bash` /bin/citbash
-             fi
            '''
+
+	// built-in runs as Jenkins user so can't write to /bin (see also above)
+	if (agentName != 'built-in') {
+	    sh '''
+                 if [ ! -f /bin/citbash ]; then
+                   ln -sf `which bash` /bin/citbash
+                 fi
+               '''
+	}
     }
 }
